findlib: Create a bbclass to handle ocamlfind.
Ocaml library manager.  Use a local findlib.conf to handle ocamlfind
commands in recipes sysroots.

Make a cross recipe for findlib to use with the ocaml cross-toolchain.
ocamlfind will use META configuration files installed with the libraries
to figure out dependencies. Although it could possibly use the native
recipe and be configured to probe the cross build environment, it is
easier to handle as a cross tool to begin with and avoid bitbake
headache with native staging directory being corrupted.

Introduce OPN, BPN without eventual "ocaml-" prefix. site-lib directory
for ocamlfind is populated predictably, so factorise that in the
findlib.bbclass.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1253