ocaml: absolute path in shebang overflow.
ocamlc bytecode will embbed a shebang pointing to the ocamlrun
interpreter. When ocaml-native is built with Bitbake in a container, the
absolute path to ocamlrun easily exhaust the common 127 character limit
of the shebang.

Bitbake will setup the environment to handle
hosttools/sysroot-native/sysroot invocations, so it makes sense to use
env in the shebang instead of an absolute path and rely on the
environment provided by Bitbake to find the appropriate ocamlrun
interpreter, and avoid shebang overflow.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1279