ocaml: local findlib package database
Findlib configuration should be part of the sysroot. This prevents
operations in the version control system to wipe it without Bitbake
knowledge resulting in build failures (e.g, externalsrc).

Some entries require absolute path, so some substitution is required.
Since the recipe paths are required are relative to the recipe,
do_populate_sysroot[postfuncs] is not possible, instead add a
prepare_recipe_sysroot[postfuncs] in the class to handle the
the changes.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>