ocaml: local findlib package database
The database should be in ${B}, no in ${S}.

If externalsrc is inherited and a local tree is used, since this recipe
is MACHINE specific, it might find the configuration intended for
another MACHINE in its sources and used it leading to all sorts of
problems.

do_{fetch,unpack,patch} are not defined with externalsrc.bbclass, use
do_prepare_recipe_sysroot. It is more accurate since the configuration
file is only generated from environment variables and static path.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>