ocaml/findlib: fix toolchain cross environment
Ocaml projects mostly use a common set of variables to find the native
and byte-code compiler. These in turn may use the appropriate C compiler
and libraries to interface with C sources and produce useable binaries.

Use the common OCAMLC, OCAMLOPT and OCAMLMKLIB to pass the appropriate
options from the ocaml bbclass. Also ignore the default ldconf in the
findlib bbclass as this will be provided through OE.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1787